Set timeout to 0 for advanced ui testing
That commit set weddrier.implicity_timeout to 0.

For some cases working without implicity timeout lead
to troubles with founding something (we are faster than browser).

In that cases it is better to repeat some operation instead of
setting to timeout everything.

I made special decorator, that allow to use it against search
function of driver to repeat a search.