ADCM-1366 MultipleObjectsReturned on GET hc
There is a trouble with read operation of HC in cluster.
That is an issue in ADCM-733 solution.

In case we have more than one prototype of service with the
same name in DB we have a trouble with searching exact component.
There is no restriction per bundle. So we have MultipleObj
exception on objects.get() operation.

Solution is to restrict service prototypes to exactly one bundle.

That solution is not simple because we have no good ORM way to find
result over FK. We have to use SQL in operation, which not so offen in
use in our code.