Simplify url config patterns with include
There is a way to reuse url and view names. See documentation at
https://docs.djangoproject.com/en/3.1/topics/http/urls/#passing-extra-options-to-include

In that sample I use the same patterns and view for two objects:
ADCM and Provider. I beleave it is possible to do so for Cluster,
Service and Host too.

After changes I checked url on API and schema over client. Everything
is fine.