ADCM-2420 Ansible 2.9 as alternative
The commit adds an alternative variant to run playbooks. From now
a bundle developer could set a venv atts on action or on object
to point to a custom venv.

From now we have two venv:

default - with Ansible 2.8.x
2.9 - with Ansible 2.9.x

If venv is ommited in config.yaml, then we have "default" venv.