ADCM-5521 Altered APIClient class for more obvious path building
Motivation

`reverse` was mostly designed for non-API Django for cases when paths change, which isn't a normal thing for modern APIs.
`reverse` requires **a lot** of space, especially for nested endpoints and remembering all views' names is quite inconvenient.

Solution

Simple path resolver, so it'll be clear path for which object is built.
Current implementation is what I see as balance between "obvious" and "convenient" both to read and write.